# backend/internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Palhelm API", "version": "1.0.0"},
  "security": [{"session": []}],
  "paths": {
...
    "/api/integration/v1/server": {
      "get": {"summary": "Server status from the poller's cached last-successful snapshot (never a per-request upstream call)", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IntegrationIfNoneMatch"}], "responses": {
        "200": {"description": "Server status; state is \"unreachable\" with zero-value fields when no snapshot exists, never a 5xx", "headers": {"ETag": {"$ref": "#/components/headers/ETag"}, "Cache-Control": {"$ref": "#/components/headers/NoStore"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IntegrationServerEnvelope"}}}},
        "304": {"$ref": "#/components/responses/IntegrationNotModified"},
        "401": {"$ref": "#/components/responses/IntegrationUnauthorized"},
        "429": {"$ref": "#/components/responses/IntegrationRateLimited"}
      }},
      "post": {"summary": "Only GET is registered on this router; every other method is rejected after auth", "security": [{"bearerAuth": []}], "responses": {"405": {"$ref": "#/components/responses/IntegrationMethodNotAllowed"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {"type": "apiKey", "in": "cookie", "name": "palhelm_session"},
      "bearerAuth": {"type": "http", "scheme": "bearer", "description": "Integration API key, format phk_<8 lowercase hex>_<43 base64url chars> (56 chars total). Created via POST /api/v1/integration-keys; the plaintext is shown exactly once. Scoped to the GET-only /api/integration/v1 group and never accepted by any session-gated route. Treat keys as passwords and serve this API over TLS at the network edge."}
    },
    "parameters": {
...
      "IntegrationIfNoneMatch": {"name": "If-None-Match", "in": "header", "required": false, "description": "Comma-separated list of ETags (weak comparison) or \"*\"; a match returns 304 with no body.", "schema": {"type": "string"}},
...
    },
    "headers": {
      "ETag": {"description": "Weak content-hash ETag (SHA-256 of the response body, first 32 hex chars, W/-prefixed). Present on every integration 200 and echoed unchanged on the resulting 304; never present on 4xx/5xx.", "schema": {"type": "string"}},
      "NoStore": {"description": "Always \"no-store\": set on every /api/integration/v1 response regardless of status, and on every /api/v1/integration-keys response.", "schema": {"type": "string", "const": "no-store"}},
      "RetryAfter": {"description": "Seconds to wait before retrying. Present on every 429. On the fail-closed limiter-map-full path this is the full 60s window, since there is no oldest timestamp to compute a tighter value from.", "schema": {"type": "integer", "minimum": 1}},
      "WWWAuthenticate": {"description": "Present on every 401 from the integration group.", "schema": {"type": "string", "const": "Bearer"}},
      "Allow": {"description": "Present on every 405 from the integration group.", "schema": {"type": "string", "const": "GET"}}
    },
    "responses": {
      "IntegrationUnauthorized": {"description": "Missing, duplicate, malformed, unknown-id, wrong-secret, or revoked bearer token. Every failure mode returns this identical body and headers - there is no distinguishable variant to probe with.", "headers": {"WWW-Authenticate": {"$ref": "#/components/headers/WWWAuthenticate"}, "Cache-Control": {"$ref": "#/components/headers/NoStore"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "IntegrationRateLimited": {"description": "The per-key rate limit (default 60 req/min, PALHELM_INTEGRATION_RATE_LIMIT) was exceeded.", "headers": {"Retry-After": {"$ref": "#/components/headers/RetryAfter"}, "Cache-Control": {"$ref": "#/components/headers/NoStore"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "IntegrationMethodNotAllowed": {"description": "The integration router registers only GET handlers; every other method on a real or unknown path is rejected after auth with this envelope.", "headers": {"Allow": {"$ref": "#/components/headers/Allow"}, "Cache-Control": {"$ref": "#/components/headers/NoStore"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "IntegrationNotModified": {"description": "If-None-Match matched the current representation (weak comparison, member-wise, or a bare \"*\"). No body.", "headers": {"ETag": {"$ref": "#/components/headers/ETag"}, "Cache-Control": {"$ref": "#/components/headers/NoStore"}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message"], "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "manualCommand": {"type": "string"}}}}},
...
      "IntegrationSave": {"type": "object", "required": ["state", "formatDrift", "lastParseAt", "players", "pals", "guilds"], "properties": {"state": {"type": "string", "enum": ["ok", "drift", "unknown"]}, "formatDrift": {"type": "boolean"}, "lastParseAt": {"type": ["string", "null"], "format": "date-time"}, "players": {"type": "integer"}, "pals": {"type": "integer"}, "guilds": {"type": "integer"}}},
      "IntegrationServer": {"type": "object", "required": ["name", "description", "version", "state", "uptimeSec", "save"], "properties": {"name": {"type": "string"}, "description": {"type": "string"}, "version": {"type": "string"}, "state": {"type": "string", "enum": ["running", "countdown", "stopping", "unreachable"]}, "uptimeSec": {"type": "integer", "format": "int64"}, "save": {"$ref": "#/components/schemas/IntegrationSave"}}},
...
      "IntegrationServerEnvelope": {"type": "object", "required": ["data"], "properties": {"data": {"$ref": "#/components/schemas/IntegrationServer"}}},
...
    }
  }
}